+++ watermelon/administration/src/main/java/xyz/nanian/owl/admin/service/SpotlightService.java
@@ -17,9 +17,9 @@ import java.util.List;
  */
 public interface SpotlightService extends IService<SpotlightDO> {
     List<SpotlightVO> listByOrder();
-    SpotlightVO getById(Integer id);
+    SpotlightVO getById(Long id);
     int create(SpotlightDTO dto);
     Boolean update(SpotlightDTO dto);
-    Boolean deleteById(Integer id);
+    Boolean deleteById(Long id);
 
 }
